refactor: componentize minimal-auth-page.tsx  into modular auth components
