+++ components/auth/AuthBackground.tsx
@@ -1,0 +1,25 @@
+'use client';
+
+import React from 'react';
+import { Particles } from '@/components/ui/particles';
+
+export function AuthBackground() {
+	return (
+		<>
+			<Particles
+				color="#f44545"
+				quantity={120}
+				ease={20}
+				className="absolute inset-0"
+			/>
+			<div
+				aria-hidden
+				className="absolute inset-0 isolate -z-10 contain-strict"
+			>
+				<div className="bg-[radial-gradient(68.54%_68.72%_at_55.02%_31.46%,--theme(--color-foreground/.06)_0,hsla(0,0%,55%,.02)_50%,--theme(--color-foreground/.01)_80%)] absolute top-0 left-0 h-320 w-140 -translate-y-87.5 -rotate-45 rounded-full" />
+				<div className="bg-[radial-gradient(50%_50%_at_50%_50%,--theme(--color-foreground/.04)_0,--theme(--color-foreground/.01)_80%,transparent_100%)] absolute top-0 left-0 h-320 w-60 [translate:5%_-50%] -rotate-45 rounded-full" />
+				<div className="bg-[radial-gradient(50%_50%_at_50%_50%,--theme(--color-foreground/.04)_0,--theme(--color-foreground/.01)_80%,transparent_100%)] absolute top-0 left-0 h-320 w-60 -translate-y-87.5 -rotate-45 rounded-full" />
+			</div>
+		</>
+	);
+}
