+++ components/auth/index.ts
@@ -1,0 +1,7 @@
+export { AuthBackground } from './AuthBackground';
+export { AuthNavigation } from './AuthNavigation';
+export { AuthLogo } from './AuthLogo';
+export { AuthHeader } from './AuthHeader';
+export { AuthButtons } from './AuthButtons';
+export { AuthPromoBanner } from './AuthPromoBanner';
+export { AuthFooter } from './AuthFooter';
